% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm-class.R
\name{Amat}
\alias{Amat}
\title{Extract a projection matrix from an \code{mpm} object}
\usage{
Amat(x, census = attr(x, "census"))
}
\arguments{
\item{x}{An \code{mpm} object.}

\item{census}{A string specifying the census to use. Defaults to the value
encoded in \code{x}.}
}
\value{
A projection matrix, which is a matrix object with row and column
names.
}
\description{
\code{Amat} extracts the projection matrix (often called the "A matrix") from a
matrix population model (\code{mpm}) object. By default, the census coded into the
mpm object is used, but this can be overridden by specifying a value for
\code{census}.
}
\examples{
# Make a Leslie matrix model with postbreeding census
bx <- c(0, 0.040, 1.470, 2.065, 2.440, 3.250, 3.250, 3.250)     # Births
px <- c(0.424, 0.726, 0.513, 0.361, 0.175, 0.700, 0.286, 0)     # Survival
my_lifetable <- data.frame(age = 0:7, mx = bx, Px = px)
my_mpm <- mpm(my_lifetable, "Leslie")

# Extract the projection matrices
Amat(my_mpm)            # Postbreeding census
Amat(my_mpm, "pre")     # Prebreeding census
}
