% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Caswell_Ex_2.1}
\alias{Caswell_Ex_2.1}
\title{Example 2.1 from Caswell (2001)}
\format{
A matrix with 3 rows and 3 columns
}
\source{
Caswell, H. 2001. Matrix population models: Construction, analysis, and
interpretation. 2nd edition. Sinauer Associates, Sunderland, MA.
}
\usage{
Caswell_Ex_2.1
}
\description{
A 3 x 3 Leslie matrix model. Given as example 2.1, equation (2.8).
}
\keyword{datasets}
