% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_stage.vector}
\alias{collapse_stage.vector}
\title{Reduce number of stages in MPM simulation output}
\usage{
collapse_stage.vector(stage.vectors, stages = NULL)
}
\arguments{
\item{stage.vectors}{A matrix, with each row being a class in the MPM and each column being a time
step. The stage.vectors element of pop.projection satisfies this.}

\item{stages}{An optional vector, specifying the stage associated with each class of the MPM. Should
be the same length as the number of rows of stage.vectors. If the matrix was produced using a Leslie matrix created using
\code{make_stage4age_matrix}, then this can be left unspecified.}
}
\value{
A matrix similar to the input stage.vectors, but with the rows collapsed into the reduced number of stages.
}
\description{
This is designed to make it easy to collapse simulation output from an age-structured
model into stages for ease in plotting. It works well with the output provided by pop.projection
in the popbio package.
}
\examples{
# Leslie matrix version of the same data of the loggerhead population
A2 <- make_stage4age_matrix(loggerhead[[2]])
N0 <- rep(10, nrow(A2))
library(popbio)
pop.sim <- pop.projection(A2, N0)
sv <- collapse_stage.vector(pop.sim$stage.vector)
stage.vector.plot(sv)
}
