% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen-analysis.R
\name{lambda1}
\alias{lambda1}
\title{Asymptotic growth rate of an MPM.}
\usage{
lambda1(A, census = NULL)
}
\arguments{
\item{A}{Either an mpm object or a square matrix representing a matrix
population model}

\item{census}{A string giving the census to use when extracting the
projection matrix from A (if it's an mpm object). If not specified, the
default census coded into A is used. Value is ignored if A is a matrix.}
}
\value{
The dominant eigenvalue of \code{A}, which is the asymptotic growth rate of the
deterministic MPM represented by A.
}
\description{
Calculates lambda_1, the dominant eigenvalue of a matrix, which can be interpreted as
the asmyptotic population growth rate of a matrix population model.
}
\examples{
lambda1(Caswell_Ex_2.1)
}
