% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loggerhead}
\alias{loggerhead}
\title{Loggerhead sea turtle stage-based life tables}
\format{A list with two elements, each being a data frame with four named columns.}
\source{
\code{loggerhead[[1]]} is derived from Table 3 of Crouse et al. (1987), and
\code{loggerhead[[2]]} is Table 1 of Crowder et al. (1994).
}
\usage{
loggerhead
}
\description{
Two stage-based life tables (differing in how adults are broken out) of the Little Cumberland Island loggerhead sea turtle
population, set up to use as
input to \code{\link{make_stage4age_matrix}}. Note that these tables do not exactly
generate the stage-based matrices published in the sources. The matrix in Table 4 of
Crouse et al. (1987) has inappropriate reproduction matrix elements for a
postbreeding-census matrix, and has an apparent typo in the survival of adults. The
matrix in Table 2 of Crowder et al. (1994) differs in the third decimal place of
subadult reproduction, probably as a consequence of rounding the maturation fraction
before multiplying by the adult maternity.
}
\references{
Crouse, D. T., L. B. Crowder, and H. Caswell. 1987. A stage-based
population model for loggerhead sea turtles and implications for conservation.
Ecology 68:1412-1423.

Crowder, L. B., D. T. Crouse, S. S. Heppell, and T. H. Martin. 1994. Predicting the
impact of turtle excluder devices on loggerhead sea turtle populations. Ecological
Applications 4:437-445.
}
\keyword{datasets}
