% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm_creation.R
\name{make_Lefkovitch_matrix}
\alias{make_Lefkovitch_matrix}
\title{Create a Lefkovitch stage-structured matrix}
\usage{
make_Lefkovitch_matrix(
  stage_table,
  survival = stage_table$survival,
  maternity = stage_table$maternity,
  maturation = stage_table$maturation,
  model = c("post", "pre")
)
}
\arguments{
\item{stage_table}{Either a vector of stage names or a data frame with columns
"stage_name", "survival", "maternity", and "maturation". In the latter case the next
three arguments do not need to be provided}

\item{survival}{A vector of stage-specific survival, on a per-timestep basis}

\item{maternity}{A vector of stage-specific maternities (number of offspring produced
by an individual in a given stage)}

\item{maturation}{A vector of the maturation rate, g_i, representing the fraction of
individuals that (conditional on survival) should move on to the next stage. If the
stage only lasts one timestep (all surviving individuals mature) then use a value
of 1.}

\item{model}{Whether the matrix should represent a prebreeding ("pre") or postbreeding
("post") census model. Defaults to "post"}
}
\value{
A matrix representing the Lefkovitch MPM.
}
\description{
In a Lefkovitch model, individuals may either stay in the current stage or transition
to the next stage, with the rate of moving from stage i to stage i+1, conditional on
survival, being given as g_i (commonly called "growth" but really representing
"maturation"). This function creates a Lefkovitch matrix using user-supplied
information on maturation rates as well as stage-specific survival and maternity
rates.
}
