% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm_creation.R
\name{make_Leslie_matrix}
\alias{make_Leslie_matrix}
\title{Create a Leslie Matrix Model}
\usage{
make_Leslie_matrix(x, sx = NULL, mx = NULL, last = c("repeating",
  "terminal"), model = c("pre", "post"))
}
\arguments{
\item{x}{Either (1) a data frame containing columns named "x" (age, starting with age
zero), "sx" (containing one-timestep survival rates for each age) and "mx"
(containing birth rates associated with each age); or (2) an integer giving the
maximum age (in timesteps) for the model; or (3) a vector of ages.}

\item{sx}{A vector of age-specific survival rates. Not needed if x is a data frame}

\item{mx}{A vector of age-specific birth rates Not needed if x is a data frame}

\item{last}{One of "repeating" (the last age class has a survival self-loop) or
"terminal" (there is no survival from the final age class)}

\item{model}{The type of model. Currently supported options are "pre" (prebreeding
census model, the default) and "post" (postbreeding census model)}
}
\value{
A matrix containing the Leslie matrix model
}
\description{
Create a Leslie matrix from a schedule of age-specific survival and birth rates.
The age classes must be evenly spaced, and survival is from one age class to the next.
The final age class may be terminal (no further survival) or can be constructed with
a self-loop for indefinite age classes having the same demography as the final one.
}
\details{
If the model is terminal, then the survival of the last age class must be zero.
}
\examples{
# Create a demography schedule, with juvenile and senescent age classes
demog_sched <- data.frame(x = 0:7,
                          sx = c(0.05, 0.2, 0.35, 0.8, 0.9, 0.9, 0.75, 0.4),
                          mx = c(0, 0, 0, 0.5, 1, 3, 3, 1.5))

# Make a Leslie matrix using the data frame and the default model structures
make_Leslie_matrix(demog_sched)

# Supply the vectors directly, and get a postbreeding census model
with(demog_sched, make_Leslie_matrix(0:8, c(sx, 0), c(mx, 1), model = "post"))

# Supply x as an integer, and get a terminal model ("true" Leslie matrix)
with(demog_sched, make_Leslie_matrix(7, sx, mx, last = "terminal"))
}
