% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm_creation.R
\name{make_stage4age_matrix}
\alias{make_stage4age_matrix}
\title{Construct a stage-structured matrix with given mean stage durations}
\usage{
make_stage4age_matrix(stage_table, survival = stage_table$survival,
  maternity = stage_table$maternity, duration = stage_table$duration,
  approx_method = c("unrolled", "AAS", "SAS", "FAS"), model = c("post",
  "pre"))
}
\arguments{
\item{stage_table}{Either a vector of stage names or a data frame with columns
"stage_name", "survival", "maternity", and "duration". In the latter case the next
three arguments do not need to be provided}

\item{survival}{A vector of stage-specific survival, on a per-timestep basis}

\item{maternity}{A vector of stage-specific maternities (number of offspring produced
by an individual in a given stage)}

\item{duration}{A vector of average number of timesteps spent in each stage. If the
final stage continues indefinitely (there is no maximum age) then the last element
should be "Inf"}

\item{approx_method}{The rule for generating the matrix (see details). Defaults to
"unrolled"}

\item{model}{Whether the matrix should represent a prebreeding ("pre") or postbreeding
("post") census model. Defaults to "post"}
}
\value{
A projection matrix. For the "unrolled" case the row and column names are a
concatenation of the stage name and the age; for the others they are the stage names.
}
\description{
Construct a stage-structured matrix with given mean stage durations
}
\details{
There is no universally "best" way to construct a stage structured model
based on mean stage durations. This function implements four approaches, with names
based on the sheme in Kendall et al. (in review).

\code{approx_method = "unrolled"}: this creates an age-structured Leslie matrix where stage i
is replicated \code{duration[i]} times. This is a good solution if the variance in stage
duration is small, and is the best solution if there is no variance in stage duration:
it correctly generates both the transient and asymptotic dynamics. The cost is a
potentially large matrix, which doesn't cause R any difficulty but may be challenging
to visualize (future developments will provide tools to aggregate results by stage).
This method requires that all elements of \code{duration} be integers.
}
