% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpm_creation.R
\name{pre_to_post}
\alias{pre_to_post}
\title{Convert prebreeding census MPM to corresponding postbreeding census MPM}
\usage{
pre_to_post(S0, Amat = NULL, Fmat = NULL, Umat = NULL)
}
\arguments{
\item{S0}{Newborn survival. If Fmat is not supplied, or has only one non-zero row, then
S0 should be a scalar. If Fmat has more than one non-zero row, then S0 may be a
scalar (in which case all newborn classes have the same survival) or a vector with
length equalling the number of non-zero rows in Fmat.}

\item{Amat}{A square matrix representing a prebreeding census MPM}

\item{Fmat}{A matrix the same size of Amat, containing positive elements for the
fertility coefficients, and zeros elsewhere}

\item{Umat}{A matrix the same size of Amat, containing positive elements for the
survival/transition coefficients, and zeros elsewhere}
}
\value{
A matrix containing the postbreeding census MPM that corresponds to Amat (or to
  Fmat+Umat). Its rank will be one larger than that of Amat
}
\description{
Prebreeding census MPMs and postbreeding census MPMs are functionally equivalent (they
give the same results when analyzed), as the only difference is when the population is
counted. It is far easier to construct a prebreeding census MPM without inadvertently
introducing errors. However, such a model lacks newborns in the population vector;
users may want to include the newborns for ease of comparing the model projections with
actual censuses. This function automates the conversion of a prebreeding into a
postbreeding census MPM.
}
\details{
S0 must be provided. In addition, provide either Amat alone (in which case the top row
of Amat will be assumed to comprise fertility coefficients), Amat and Fmat, or Fmat and
Umat.

Currently, the only matrix structure supported is one in which only the first row of
the matrix contains fertility coefficients. This is particularly suitable for
non-spatial, one-sex or asexual models of animals (or plants without a seedbank).
}
\section{To do}{

\itemize{
  \item Account for multiple non-zero rows of Fmat
  \item Allow one newborn type to grow to multiple classes at age one
  \item Allow multiple newborn types
  \item Add argument testing
  \item Allow Amat to be an mpm object
}
}

