% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeasonalDemography.R
\docType{class}
\name{SeasonalDemography-class}
\alias{SeasonalDemography-class}
\alias{SeasonalDemography}
\title{Create a representation of structured demography over a single season}
\usage{
SeasonalDemography(
  class_names,
  season,
  class_type,
  newborn_classes,
  survival,
  maternity,
  transition
)
}
\value{
An object of class \code{SeasonalDemogaphy}
}
\description{
The \code{SeasonalDemography} class contains information about the survival,
maternity, and class transitions (aging, growth, maturation) of a structured
population over a single season. The season may be an actual part of the
calendar, or it may be a conceptual part of the annual cycle (e.g.,
"breeding" or "nonbreeding" for a simple birth-pulse model, or "northward
migration" for a migratory population). It is a building block of the
\code{Demography} class, and many users will not need to access it directly.
}
\details{
The constructor function requires that all slots be specified, and performs
basic validity checking on dimensions and vector lengths.

At present, there are no functions for extracting or replacing the values in
indvidual slots in an existing \code{SeasonalDemography} object. The slots can, of
course, be accessed directly, but there will be no validity checking.
}
\section{Slots}{

\describe{
\item{\code{class_names}}{A character vector (or something that can be coerced to
one) giving the names of classes (ages, stages, etc.) of the population.
Age classes may be provided as a numeric vector; it will be silently
converted to character.}

\item{\code{season}}{The name of the season. May be any text string.}

\item{\code{class_type}}{The type of structuring variable (e.g., "age", "stage",
"age-size"). May be any text string.}

\item{\code{newborn_classes}}{The name(s) of the class(es) that represent newborn
individuals. If there is only one such class, a string; otherwose a
character vector. All elements of \code{newborn_classes} must be found in
\code{class_names}.}

\item{\code{survival}}{A numeric vector giving the survival probabilities of each
class over the entire season.}

\item{\code{maternity}}{A numeric matrix giving the number of newborns in each class
produced by each of the non-newborn classes over the course of the season.
There should be a row for each newborn class and a column for each
non-newborn class. Newborns are assumed to have zero maternity; if
\code{ncol(maternity) == length(class_names)}, then the newborn classes will be
dropped. If there is a single newborn class, \code{maternity} may be provided as
a numeric vector.}

\item{\code{transition}}{A numeric matrix giving the class transitions. The value in
\code{transition[i, j]} gives the probability (conditional on survival) that an
individual starting the season in class \code{j} ends the season in class \code{i}.
Because the transitions are conditional on survival, columns should sum to
one.}
}}

\examples{
SeasonalDemography(c("Newborn", "Juvenile", "Adult"),
                   season = "nonbreeding",
                   class_type = "stage",
                   newborn_classes = "Newborn",
                   survival = c(0.1, 0.5, 0.9),
                   maternity = rep(0, 3),
                   transition = matrix(c(0, 0, 0, 1, 0.8, 0, 0, 0.2, 1),
                                       nrow = 3, byrow = TRUE))
}
