% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen-analysis.R
\name{stable_stage}
\alias{stable_stage}
\title{Stable stage distribution of an MPM}
\usage{
stable_stage(A, collapse = FALSE, stages = NULL, census = NULL)
}
\arguments{
\item{A}{Either an mpm object or a square matrix representing a matrix
population model}

\item{collapse}{Should classes in the matrix be collapsed into a smaller
number of stage classes?}

\item{stages}{A vector with length equal to the rank of A giving the stage
name for each matrix class. If A is a Leslie matrix created using
\code{make_stage4age_matrix}, then this can be left unspecified.}

\item{census}{A string giving the census to use when extracting the
projection matrix from A (if it's an mpm object). If not specified, the
default census coded into A is used. Value is ignored if A is a matrix.}
}
\value{
A vector of the stable stage distribution
}
\description{
Calculates w_1, the eigenvector associated with the dominant eigenvector,
which can be interpreted as the stable age/stage/size distribution. The
vector is scaled to sum to one. Allows stage-for-age Leslie matrices to be
reported by stage for ease of visualization.
}
\examples{
# Reproduce the model in Crowder et al:
A1 <- make_stage4age_matrix(loggerhead[[2]], approx_method = "AAS")
stable_stage(A1)

# Now do the Leslie matrix version of the same data
A2 <- make_stage4age_matrix(loggerhead[[2]])
stable_stage(A2) # Age structure
stable_stage(A2, collapse = TRUE) # Stage structure
}
