% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subdiag}
\alias{subdiag}
\title{Create Subdiagonal Matrix}
\usage{
subdiag(A, sx)
}
\arguments{
\item{A}{Either a positive integer giving the row and column dimensions of the
matrix, or a square matrix}

\item{sx}{Values to put on the subdiagonal. Either a scalar (which will be
replicated as needed) or a vector with length one less than \code{A} or \code{nrow(A)} (as
appropriate)}
}
\value{
A matrix with specified subdiagonal or with the subdiagonal added to the
input matrix
}
\description{
Create a square matrix with specified subdiagonal elements, or add specified elements
to the subdiagonal of an existing matrix.
}
\details{
If \code{A} is an integer \eqn{n}, then \code{subdiag(A, sx)} returns an \eqn{n \times
n}{n x n} zero matrix with nonzero values on the subdiagonal. If \code{A} is a square
matrix, then \code{subdiag(A, sx)} returns the matrix \code{A} with the subdiagonal
elements added to it.
}
\examples{
# Create zero matrix with non-zero subdiagonal
subdiag(5, 1:4)

# Add subdiagonal elements to existing matrix
A <- matrix(1:16, 4, 4, byrow = TRUE)
subdiag(A, 0.5)

# Create zero matrix with non-zero subdiagonal that is the size of A
subdiag(nrow(A), 0.5)
}
