% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teasel}
\alias{teasel}
\title{Teasel demography}
\format{
A list with two elements:
\describe{
\item{demog}{A data frame with three columns: \code{stage}, \code{seed_production},
and \code{survival}. These give the stage names, the stage-specific seed
production, and stage-specific survival, respectively.}
\item{transitions}{A 7x7 numeric matrix giving the transition
probabilitys (conditional upon survival) between stages.}
}
}
\source{
Werner & Caswell (1977), Table 3.
}
\usage{
teasel
}
\description{
Reproduction, survival, and growth transitions for an experimental population
of teasel (\emph{Dipsacus sylvestris}). This is the "Field A" population of Werner
& Caswell (1977). Note that the matrix in that paper is incorrect; Caswell
(2001) gives the correct form for a prebreeding-census model.
}
\references{
Caswell, H. 2001. Matrix population models: Construction,
analysis, and interpretation. 2nd edition. Sinauer Associates, Sunderland,
MA.

Werner, P. A., and H. Caswell. 1977. Population growth rates and age versus
stage-distribution models for teasel (\emph{Dipsacus sylvestris} Huds.). Ecology
58:1103-1111.
}
\keyword{datasets}
