% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen-analysis.R
\name{lambda1}
\alias{lambda1}
\title{Asymptotic growth rate of an MPM.}
\usage{
lambda1(A)
}
\arguments{
\item{A}{A square matrix}
}
\value{
The dominant eigenvalue of `A`, which is the asymptotic growth rate of the
  deterministic MPM represented by A.
}
\description{
Calculates lambda_1, the dominant eigenvalue of a matrix, which can be interpreted as
the asmyptotic population growth rate of a matrix population model.
}
\examples{
lambda1(Caswell_Ex_2.1)
}
